+++ src/reports/reports.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { ReportsService } from './reports.service';
+import { ReportsController } from './reports.controller';
+import { PrismaService } from '@app/prisma/prisma.service';
+import { AuthModule } from '@app/auth/auth.module';
+
+@Module({
+  imports: [AuthModule],
+  controllers: [ReportsController],
+  providers: [ReportsService, PrismaService],
+})
+export class ReportsModule {}